Re-formatted and rearranged get method in java files
