+++ src/main/java/com/google/ehub/servlets/LoginServlet.java
@@ -39,6 +39,7 @@ public class LoginServlet extends HttpServlet {
       response.getWriter().println("<p>Login <a href=\"" + loginUrl + "\">here</a>.</p>");
     }
   }
+  
   /**
    * Returns the email of the currently logged in user
    */
