+++ src/main/java/com/google/ehub/servlets/ProfileServlet.java
@@ -8,7 +8,6 @@ import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.ehub.data.ProfileDatastore;
-import com.google.ehub.data.UserDataManager;
 import com.google.ehub.data.UserProfile;
 import com.google.gson.Gson;
 import java.io.IOException;
@@ -30,10 +29,11 @@ public class ProfileServlet extends HttpServlet {
   private static final String EMAIL_PROPERTY_KEY = "email";
   private static final String USERNAME_PROPERTY_KEY = "username";
   private static final String BIO_PROPERTY_KEY = "bio";
+  ProfileDatastore profileData = new ProfileDatastore();
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    ProfileDatastore profileData = new ProfileDatastore();
+    
 
     String name = request.getParameter(NAME_PROPERTY_KEY);
     String email = request.getParameter(EMAIL_PROPERTY_KEY);
@@ -51,14 +51,14 @@ public class ProfileServlet extends HttpServlet {
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    UserDataManager manager = new UserDataManager();
     LoginServlet login = new LoginServlet();
     String useremail = login.getEmail();
-    UserProfile newUser = manager.getUserProfile(useremail);
+    UserProfile newUser = profileData.getUserProfile(useremail);
 
     response.setContentType("application/json");
     response.getWriter().println(convertToJson(newUser));
   }
+
   /**
    * Creates a json from the UserProfile object.
    *
@@ -68,6 +68,7 @@ public class ProfileServlet extends HttpServlet {
   private String convertToJson(UserProfile profile) {
     return new Gson().toJson(profile);
   }
+
   /**
    * Checks if any of the request values are null.
    *
